requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
gensim==4.3.1
nltk==3.8.1
zemberek-python==0.1.0
tqdm==4.65.0 